refactor(daemon): extract 4 static inner classes from StreamingAgentHandler
StreamingAgentHandler had grown to 4,555 LoC — the largest single file in
the repo. Four self-contained static inner classes inside it accounted for
~1,400 of those lines and were the obvious extraction candidates: all
static (no outer-state capture), no external type references (only comment
mentions in five tests).

## What moved

| New file | LoC | Was inner class of SAH |
|---|---|---|
| CheckpointingPlanEventListener.java |  168 | yes (lines 3169-3308) |
| StreamingNotificationHandler.java   |  255 | yes |
| CancelAwareStreamContext.java        |  492 | yes |
| PermissionAwareTool.java             |  565 | yes |
| PendingPermission.java               |   23 | private record in SAH |

`PendingPermission` is a small (3-field) record that both
`StreamingAgentHandler` and `CancelAwareStreamContext` reach for; promoting
it to top-level avoided the awkward inner-of-extracted-sibling reference.

The `TOOL_PERMISSION_LEVELS` static map (read in exactly one place —
PermissionAwareTool.execute) moved with `PermissionAwareTool`, so "what
default permission level does each tool have?" lives next to the wrapper
that asks the question.

## Result

StreamingAgentHandler: **4,555 → 3,214 LoC (-29%)**

Diff overall: +1,511 / -1,351 = **+160 net** lines (per-file headers,
imports, and now-explicit class-level javadocs that previously rode on
the inner-class hosts).

## Visibility bumps

Three things became package-private (were private when same-class):

- `CancelAwareStreamContext.PERMISSION_RESPONSE_TIMEOUT_MS` (constant read
  by `PermissionAwareTool.execute` for the awaiting-approval timeout)
- `StreamingAgentHandler.AntiPatternGateOverrideStatus` (public nested
  record — left in SAH, referenced from PermissionAwareTool by qualified
  name; no API surface change)
- Nothing else — the rest of the moved code was either already package-
  private or the moved methods don't cross any access boundary.

No `public` API surface added or removed. No call sites changed semantics.

## Test fix

`PermissionResponseRoutingTest.setUp` used reflection to find
`PendingPermission` via `StreamingAgentHandler.class.getDeclaredClasses()`.
After the extraction, the type is `dev.aceclaw.daemon.PendingPermission` —
load it directly via `Class.forName` instead. Same instantiation, same
field-population behaviour — only the discovery path changed.

The test's comment about "private PendingPermission record" was updated
to "package-private" to match the new visibility.

## Test plan

- [x] :aceclaw-daemon:compileJava — clean
- [x] :aceclaw-daemon:compileTestJava — clean
- [x] DaemonIntegrationTest (25 tests, 1 pre-existing flake on
  testAdaptiveReplan_revisesAndCompletes verified identical on main)
- [x] PermissionRoutingTest (8 / 0)
- [x] PermissionResponseRoutingTest (13 / 0 — was 13 failures from the
  reflection-on-inner-class issue; fix verified)
- [x] HookIntegrationTest (4 / 0)
- [x] ToolPermissionRouterTest (7 / 0)
- [x] SubAgentStructuralProbeTest (7 / 0)
- [x] Local full daemon test rerun: 0 failures

## What this is NOT

Not a behaviour change. Not a redesign of the permission flow or streaming
notification flow. Not a new abstraction. The four classes do exactly what
they did before — at a top level in the same package, with explicit
package-private visibility where the same-class shortcut no longer applies.

The remaining 3,214 LoC of StreamingAgentHandler still has further
extraction targets (skill outcome recording, anti-pattern gate overrides,
journal logging, etc.) — those involve refactoring real outer-state-using
methods and are deferred to follow-ups.
