+++ aceclaw-daemon/src/main/java/dev/aceclaw/daemon/StreamingAgentHandler.java
@@ -121,27 +121,6 @@ public final class StreamingAgentHandler {
 
     private static final Logger log = LoggerFactory.getLogger(StreamingAgentHandler.class);
 
-    /** Permission level assignments for known tools. */
-    private static final Map<String, PermissionLevel> TOOL_PERMISSION_LEVELS = Map.ofEntries(
-            Map.entry("read_file", PermissionLevel.READ),
-            Map.entry("glob", PermissionLevel.READ),
-            Map.entry("grep", PermissionLevel.READ),
-            Map.entry("list_directory", PermissionLevel.READ),
-            Map.entry("web_fetch", PermissionLevel.READ),
-            Map.entry("web_search", PermissionLevel.READ),
-            Map.entry("screen_capture", PermissionLevel.READ),
-            Map.entry("memory", PermissionLevel.READ),
-            Map.entry("task", PermissionLevel.READ),
-            Map.entry("task_output", PermissionLevel.READ),
-            Map.entry("write_file", PermissionLevel.WRITE),
-            Map.entry("edit_file", PermissionLevel.WRITE),
-            Map.entry("bash", PermissionLevel.EXECUTE),
-            Map.entry("browser", PermissionLevel.EXECUTE),
-            Map.entry("applescript", PermissionLevel.EXECUTE),
-            Map.entry("cron", PermissionLevel.READ),
-            Map.entry("defer_check", PermissionLevel.READ)
-    );
-
     private final SessionManager sessionManager;
     private final StreamingAgentLoop agentLoop;
     private final ToolRegistry toolRegistry;
@@ -202,26 +181,6 @@ public final class StreamingAgentHandler {
     private final ConcurrentHashMap<String, PendingPermission> permissionRegistry =
             new ConcurrentHashMap<>();
 
-    /**
-     * Tuple stored in {@link #permissionRegistry}: the originating
-     * session and the future to complete. Used by the WS routing
-     * path to validate cross-session attempts.
-     */
-    /**
-     * Tracks a pending permission request keyed by {@code requestId} in
-     * {@link #permissionRegistry}. The {@code context} reference lets the
-     * WS-side router ({@link #routePermissionResponse}) push a
-     * {@code permission.cancelled} notification back to the originating
-     * CLI's UDS connection so a TUI prompt blocking on local stdin can
-     * dismiss itself when a browser tab has already answered (issue #437).
-     * Without that signal, the daemon-side future completes but the
-     * CLI's local {@code PermissionBridge} future stays pending until
-     * the user types y/n into a prompt that has no effect.
-     */
-    private record PendingPermission(
-            String sessionId,
-            CompletableFuture<JsonNode> future,
-            StreamContext context) {}
 
     /** Per-session turn locks for serializing main turns within a session. */
     private final ConcurrentHashMap<String, ReentrantLock> sessionTurnLocks =
@@ -3160,152 +3119,6 @@ public final class StreamingAgentHandler {
         return result;
     }
 
-    // -- Checkpointing plan event listener ------------------------------------
-
-    /**
-     * Decorates a PlanEventListener with checkpoint persistence.
-     * Writes a checkpoint after each step completion and marks final status.
-     */
-    private static final class CheckpointingPlanEventListener
-            implements SequentialPlanExecutor.PlanEventListener {
-
-        private final SequentialPlanExecutor.PlanEventListener delegate;
-        private final PlanCheckpointStore store;
-        private volatile PlanCheckpoint currentCheckpoint;
-        private final AgentSession session;
-        private final int stepIndexOffset;
-
-        CheckpointingPlanEventListener(
-                SequentialPlanExecutor.PlanEventListener delegate,
-                PlanCheckpointStore store,
-                PlanCheckpoint initialCheckpoint,
-                AgentSession session,
-                int stepIndexOffset) {
-            this.delegate = delegate;
-            this.store = store;
-            this.currentCheckpoint = initialCheckpoint;
-            this.session = session;
-            this.stepIndexOffset = stepIndexOffset;
-        }
-
-        @Override
-        public void onStepStarted(PlannedStep step, int stepIndex, int totalSteps) {
-            delegate.onStepStarted(step, stepIndex, totalSteps);
-        }
-
-        @Override
-        public void onStepCompleted(PlannedStep step, int stepIndex, StepResult result) {
-            delegate.onStepCompleted(step, stepIndex, result);
-
-            // Apply offset: stepIndex from executor is 0-based relative to the
-            // (possibly partial) plan. For resumed plans, offset maps it back to
-            // the absolute index in the original full plan.
-            int absoluteIndex = stepIndex + stepIndexOffset;
-
-            // Update and persist checkpoint
-            var updatedPlan = currentCheckpoint.plan()
-                    .withStepStatus(step.stepId(),
-                            result.success() ? StepStatus.COMPLETED : StepStatus.FAILED);
-
-            // Serialize current conversation state
-            var conversationJson = serializeSessionMessages(session);
-
-            String hint = result.success()
-                    ? "Step " + (absoluteIndex + 1) + " completed successfully"
-                    : "Step " + (absoluteIndex + 1) + " failed: "
-                            + (result.error() != null ? result.error() : "unknown");
-
-            currentCheckpoint = currentCheckpoint.withStepCompleted(
-                    absoluteIndex, result, updatedPlan, conversationJson, hint, List.of());
-
-            try {
-                store.save(currentCheckpoint);
-            } catch (Exception e) {
-                log.warn("Failed to persist plan checkpoint after step {}: {}",
-                        stepIndex + 1, e.getMessage());
-            }
-        }
-
-        @Override
-        public void onPlanCompleted(TaskPlan plan, boolean success, long totalDurationMs) {
-            delegate.onPlanCompleted(plan, success, totalDurationMs);
-
-            currentCheckpoint = currentCheckpoint.withStatus(
-                    success ? PlanCheckpoint.CheckpointStatus.COMPLETED
-                            : PlanCheckpoint.CheckpointStatus.FAILED);
-            try {
-                store.save(currentCheckpoint);
-            } catch (Exception e) {
-                log.warn("Failed to persist final plan checkpoint: {}", e.getMessage());
-            }
-        }
-
-        @Override
-        public void onPlanReplanned(TaskPlan oldPlan, TaskPlan newPlan, int attempt, String rationale) {
-            delegate.onPlanReplanned(oldPlan, newPlan, attempt, rationale);
-
-            // Update checkpoint with the new plan
-            currentCheckpoint = new PlanCheckpoint(
-                    currentCheckpoint.planId(), currentCheckpoint.sessionId(),
-                    currentCheckpoint.workspaceHash(), currentCheckpoint.originalGoal(),
-                    newPlan, currentCheckpoint.completedStepResults(),
-                    currentCheckpoint.lastCompletedStepIndex(),
-                    serializeSessionMessages(session),
-                    currentCheckpoint.status(),
-                    "Replanned (attempt " + attempt + "): " + rationale,
-                    currentCheckpoint.artifacts(),
-                    currentCheckpoint.createdAt(), Instant.now());
-            try {
-                store.save(currentCheckpoint);
-            } catch (Exception e) {
-                log.warn("Failed to persist plan checkpoint after replan: {}", e.getMessage());
-            }
-        }
-
-        @Override
-        public void onPlanEscalated(TaskPlan plan, String reason) {
-            delegate.onPlanEscalated(plan, reason);
-
-            try {
-                store.markFailed(currentCheckpoint.planId());
-            } catch (Exception e) {
-                log.warn("Failed to mark checkpoint as failed after escalation: {}", e.getMessage());
-            }
-        }
-
-        @Override
-        public void onStepFallback(PlannedStep step, int stepIndex,
-                                   String fallbackApproach, int attempt) {
-            delegate.onStepFallback(step, stepIndex, fallbackApproach, attempt);
-        }
-
-        private static List<String> serializeSessionMessages(AgentSession session) {
-            // Simple serialization: role:content for each message
-            var result = new ArrayList<String>();
-            var messages = session.messages();
-            if (messages == null) return result;
-            for (var msg : messages) {
-                switch (msg) {
-                    case AgentSession.ConversationMessage.User u ->
-                            result.add("{\"role\":\"user\",\"content\":" + escapeJson(u.content()) + "}");
-                    case AgentSession.ConversationMessage.Assistant a ->
-                            result.add("{\"role\":\"assistant\",\"content\":" + escapeJson(a.content()) + "}");
-                    case AgentSession.ConversationMessage.System s ->
-                            result.add("{\"role\":\"system\",\"content\":" + escapeJson(s.content()) + "}");
-                }
-            }
-            return result;
-        }
-
-        private static String escapeJson(String text) {
-            if (text == null) return "null";
-            return "\"" + text.replace("\\", "\\\\")
-                    .replace("\"", "\\\"")
-                    .replace("\n", "\\n")
-                    .replace("\r", "\\r")
-                    .replace("\t", "\\t") + "\"";
-        }
-    }
 
     // -- StreamEventHandler that forwards events as JSON-RPC notifications --
 
@@ -3396,1160 +3209,6 @@ public final class StreamingAgentHandler {
         bridge.broadcast(sessionId, "stream.turn_completed", tc);
     }
 
-    /**
-     * Forwards stream events from the agent loop to the client as JSON-RPC notifications.
-     */
-    private static final class StreamingNotificationHandler implements StreamEventHandler {
-
-        private final StreamContext context;
-        private final ObjectMapper objectMapper;
-        /**
-         * Plan step id (in the dashboard's composed {@code planId:step:index}
-         * form) currently in scope, or {@code null} for non-plan ReAct turns.
-         * Set by the {@link SequentialPlanExecutor.PlanEventListener} on
-         * {@code onStepStarted} and cleared on plan termination — see #485
-         * PR 3/3. Read on {@code onContentBlockStart} for ToolUse so each
-         * {@code stream.tool_use} carries an explicit parent reference,
-         * removing the dashboard's reliance on event ordering for tool-to-step
-         * attribution.
-         */
-        private final java.util.concurrent.atomic.AtomicReference<String> currentStepId
-                = new java.util.concurrent.atomic.AtomicReference<>();
-
-        StreamingNotificationHandler(StreamContext context, ObjectMapper objectMapper) {
-            this.context = context;
-            this.objectMapper = objectMapper;
-        }
-
-        /**
-         * Sets (or clears with {@code null}) the dashboard-form step id used
-         * to tag {@code stream.tool_use} notifications. Caller is responsible
-         * for setting on step start and clearing on plan termination.
-         */
-        void setCurrentStepId(String stepId) {
-            currentStepId.set(stepId);
-        }
-
-        /**
-         * Returns the currently active plan step id, or {@code null} if no
-         * plan is in progress. Exposed so {@link PermissionAwareTool} can
-         * stamp the per-call {@link dev.aceclaw.security.Provenance} with
-         * the step that triggered each capability check (#480 PR 3).
-         */
-        String getCurrentStepId() {
-            return currentStepId.get();
-        }
-
-        @Override
-        public void onThinkingDelta(StreamEvent.ThinkingDelta event) {
-            try {
-                var params = objectMapper.createObjectNode();
-                params.put("delta", event.text());
-                String stepId = currentStepId.get();
-                if (stepId != null) {
-                    params.put("parentStepId", stepId);
-                }
-                context.sendNotification("stream.thinking", params);
-            } catch (IOException e) {
-                log.warn("Failed to send thinking delta notification: {}", e.getMessage());
-            }
-        }
-
-        @Override
-        public void onTextDelta(StreamEvent.TextDelta event) {
-            try {
-                var params = objectMapper.createObjectNode();
-                params.put("delta", event.text());
-                String stepId = currentStepId.get();
-                if (stepId != null) {
-                    params.put("parentStepId", stepId);
-                }
-                context.sendNotification("stream.text", params);
-            } catch (IOException e) {
-                log.warn("Failed to send text delta notification: {}", e.getMessage());
-            }
-        }
-
-        @Override
-        public void onContentBlockStart(StreamEvent.ContentBlockStart event) {
-            if (event.block() instanceof ContentBlock.ToolUse toolUse) {
-                try {
-                    var params = objectMapper.createObjectNode();
-                    params.put("name", toolUse.name());
-                    params.put("id", toolUse.id());
-                    String summary = summarizeToolInput(toolUse.name(), toolUse.inputJson(), objectMapper);
-                    if (!summary.isBlank()) {
-                        params.put("summary", summary);
-                    }
-                    String stepId = currentStepId.get();
-                    if (stepId != null) {
-                        params.put("parentStepId", stepId);
-                    }
-                    context.sendNotification("stream.tool_use", params);
-                } catch (IOException e) {
-                    log.warn("Failed to send tool use notification: {}", e.getMessage());
-                }
-            }
-        }
-
-        @Override
-        public void onToolCompleted(String toolUseId, String toolName,
-                                    long durationMs, boolean isError, String error) {
-            try {
-                var params = objectMapper.createObjectNode();
-                params.put("id", toolUseId);
-                params.put("name", toolName);
-                params.put("durationMs", durationMs);
-                params.put("isError", isError);
-                if (error != null && !error.isBlank()) {
-                    params.put("error", truncate(error, 160));
-                }
-                context.sendNotification("stream.tool_completed", params);
-            } catch (IOException e) {
-                log.warn("Failed to send tool completed notification: {}", e.getMessage());
-            }
-        }
-
-        @Override
-        public void onHeartbeat(StreamEvent.Heartbeat event) {
-            try {
-                var params = objectMapper.createObjectNode();
-                params.put("phase", event.phase());
-                context.sendNotification("stream.heartbeat", params);
-            } catch (IOException e) {
-                log.debug("Failed to send heartbeat notification: {}", e.getMessage());
-            }
-        }
-
-        @Override
-        public void onError(StreamEvent.StreamError event) {
-            try {
-                var params = objectMapper.createObjectNode();
-                params.put("error", event.error().getMessage());
-                context.sendNotification("stream.error", params);
-            } catch (IOException e) {
-                log.warn("Failed to send error notification: {}", e.getMessage());
-            }
-        }
-
-        @Override
-        public void onCompaction(int originalTokens, int compactedTokens, String phase) {
-            try {
-                var params = objectMapper.createObjectNode();
-                params.put("originalTokens", originalTokens);
-                params.put("compactedTokens", compactedTokens);
-                params.put("phase", phase);
-                context.sendNotification("stream.compaction", params);
-            } catch (IOException e) {
-                log.warn("Failed to send compaction notification: {}", e.getMessage());
-            }
-        }
-
-        @Override
-        public void onSubAgentStart(String agentId, String prompt) {
-            try {
-                var params = objectMapper.createObjectNode();
-                params.put("agentType", agentId);
-                params.put("prompt", prompt);
-                context.sendNotification("stream.subagent.start", params);
-            } catch (IOException e) {
-                log.warn("Failed to send subagent start notification: {}", e.getMessage());
-            }
-        }
-
-        @Override
-        public void onSubAgentEnd(String agentId) {
-            try {
-                var params = objectMapper.createObjectNode();
-                params.put("agentType", agentId);
-                context.sendNotification("stream.subagent.end", params);
-            } catch (IOException e) {
-                log.warn("Failed to send subagent end notification: {}", e.getMessage());
-            }
-        }
-
-        @Override
-        public void onUsageUpdate(long lastInputTokens, long totalInputTokens, long totalOutputTokens) {
-            try {
-                var params = objectMapper.createObjectNode();
-                params.put("inputTokens", lastInputTokens);
-                params.put("totalInputTokens", totalInputTokens);
-                params.put("totalOutputTokens", totalOutputTokens);
-                context.sendNotification("stream.usage", params);
-            } catch (IOException e) {
-                log.debug("Failed to send usage update notification: {}", e.getMessage());
-            }
-        }
-
-        private static String summarizeToolInput(String toolName, String inputJson, ObjectMapper mapper) {
-            if (inputJson == null || inputJson.isBlank()) {
-                return "";
-            }
-            try {
-                var node = mapper.readTree(inputJson);
-                if (!node.isObject()) {
-                    return truncate(inputJson, 80);
-                }
-                String value = switch (toolName) {
-                    case "bash" -> firstNonBlank(node, "command", "cmd");
-                    case "read_file", "write_file", "edit_file" -> firstNonBlank(node, "file_path", "path");
-                    case "grep" -> firstNonBlank(node, "pattern", "query");
-                    case "glob" -> firstNonBlank(node, "pattern", "path");
-                    case "list_directory" -> firstNonBlank(node, "path");
-                    case "web_fetch" -> firstNonBlank(node, "url");
-                    case "web_search" -> firstNonBlank(node, "query");
-                    case "browser" -> firstNonBlank(node, "action", "url");
-                    case "task" -> firstNonBlank(node, "prompt", "description");
-                    case "skill" -> firstNonBlank(node, "name", "skill", "prompt");
-                    default -> firstNonBlank(node, "path", "file_path", "query", "url", "command");
-                };
-                if (value == null || value.isBlank()) {
-                    return "";
-                }
-                return truncate(value.replace('\n', ' '), 80);
-            } catch (Exception ignored) {
-                return "";
-            }
-        }
-
-        private static String firstNonBlank(com.fasterxml.jackson.databind.JsonNode node, String... fields) {
-            for (var field : fields) {
-                var value = node.path(field).asText("");
-                if (!value.isBlank()) {
-                    return value;
-                }
-            }
-            return "";
-        }
-    }
-
-    // -- Cancel-aware stream context -----------------------------------------
-
-    /**
-     * A {@link StreamContext} wrapper that runs a background monitor thread
-     * to read from the socket during streaming. This is necessary because
-     * the connection thread is blocked in {@code handlePrompt()}, so
-     * {@code agent.cancel} notifications buffered in the socket go unread.
-     *
-     * <p>The monitor thread uses non-blocking NIO with a {@link Selector}
-     * to poll the socket with 100ms timeout intervals. This allows the
-     * thread to exit cleanly when stopped, without interrupting (which
-     * would close the NIO channel).
-     *
-     * <p>The monitor thread reads messages from the socket and:
-     * <ul>
-     *   <li>On {@code agent.cancel}: triggers the cancellation token</li>
-     *   <li>On {@code permission.response}: enqueues to a blocking queue
-     *       so the permission flow can consume it</li>
-     *   <li>On connection close: triggers cancellation</li>
-     * </ul>
-     */
-    private static final class CancelAwareStreamContext implements StreamContext {
-
-        private static final int READ_BUFFER_SIZE = 65536;
-        private static final long SELECT_TIMEOUT_MS = 100;
-        private static final long PERMISSION_RESPONSE_TIMEOUT_MS = 120_000;
-
-        private final StreamContext delegate;
-        private final CancellationToken cancellationToken;
-        private final ObjectMapper objectMapper;
-        private final SocketChannel channel;
-        private final StringBuilder lineBuilder;
-        private final ConcurrentHashMap<String, CompletableFuture<JsonNode>> pendingPermissions = new ConcurrentHashMap<>();
-        /**
-         * Daemon-wide registry shared with the WS bridge (issue #433) so a
-         * browser-sent {@code permission.response} can complete the same
-         * future the CLI socket monitor would. Null when running outside a
-         * full daemon (tests). All writes mirror {@link #pendingPermissions};
-         * value type carries the sessionId for the cross-session guard in
-         * {@link StreamingAgentHandler#routePermissionResponse}.
-         */
-        private final ConcurrentHashMap<String, PendingPermission> globalPermissionRegistry;
-        private final BlockingQueue<JsonNode> unmatchedResponses = new LinkedBlockingQueue<>();
-        private final Object permissionLifecycleLock = new Object();
-        private final java.util.concurrent.atomic.AtomicInteger toolUseCount =
-                new java.util.concurrent.atomic.AtomicInteger();
-        private volatile boolean stopped = false;
-        private volatile Thread monitorThread;
-        private volatile Selector selector;
-
-        CancelAwareStreamContext(StreamContext delegate, CancellationToken cancellationToken,
-                                 ObjectMapper objectMapper) {
-            this(delegate, delegate, cancellationToken, objectMapper, null);
-        }
-
-        CancelAwareStreamContext(StreamContext delegate, CancellationToken cancellationToken,
-                                 ObjectMapper objectMapper,
-                                 ConcurrentHashMap<String, PendingPermission> globalPermissionRegistry) {
-            this(delegate, delegate, cancellationToken, objectMapper, globalPermissionRegistry);
-        }
-
-        /**
-         * Two-context overload: {@code rawContext} is the original socket-backed
-         * context used for cancel-monitor channel extraction; {@code outboundContext}
-         * is what {@link #sendNotification} delegates to and is typically an
-         * {@link EventMultiplexer} that tees to the WebSocket bridge.
-         */
-        CancelAwareStreamContext(StreamContext rawContext, StreamContext outboundContext,
-                                 CancellationToken cancellationToken, ObjectMapper objectMapper) {
-            this(rawContext, outboundContext, cancellationToken, objectMapper, null);
-        }
-
-        CancelAwareStreamContext(StreamContext rawContext, StreamContext outboundContext,
-                                 CancellationToken cancellationToken, ObjectMapper objectMapper,
-                                 ConcurrentHashMap<String, PendingPermission> globalPermissionRegistry) {
-            this.delegate = outboundContext;
-            this.cancellationToken = cancellationToken;
-            this.objectMapper = objectMapper;
-            this.globalPermissionRegistry = globalPermissionRegistry;
-
-            // Extract the channel and lineBuilder from the underlying ChannelStreamContext
-            if (rawContext instanceof ConnectionBridge.ChannelStreamContext channelCtx) {
-                this.channel = channelCtx.channel();
-                this.lineBuilder = channelCtx.lineBuilder();
-            } else {
-                this.channel = null;
-                this.lineBuilder = null;
-            }
-        }
-
-        /** Number of {@code stream.tool_use} notifications observed during this request. */
-        int toolUseCount() {
-            return toolUseCount.get();
-        }
-
-        /**
-         * Registers a pending permission request keyed by requestId.
-         * Returns a future that will be completed when the matching response arrives.
-         */
-        CompletableFuture<JsonNode> registerPermissionRequest(String requestId, String sessionId) {
-            Objects.requireNonNull(requestId, "requestId");
-            Objects.requireNonNull(sessionId, "sessionId");
-            var future = new CompletableFuture<JsonNode>();
-            synchronized (permissionLifecycleLock) {
-                if (stopped) {
-                    future.cancel(false);
-                    return future;
-                }
-                var previous = pendingPermissions.putIfAbsent(requestId, future);
-                if (previous != null) {
-                    future.cancel(false);
-                    throw new IllegalStateException("Duplicate permission requestId: " + requestId);
-                }
-                // Mirror into the daemon-wide registry so the WS bridge's
-                // permission.response routing (issue #433) can complete
-                // the same future without knowing which context owns it.
-                // Tag with sessionId for the cross-session guard.
-                if (globalPermissionRegistry != null) {
-                    globalPermissionRegistry.put(requestId, new PendingPermission(sessionId, future, this));
-                }
-            }
-            return future;
-        }
-
-        /**
-         * Unregisters and cancels a pending permission request if it has not been completed.
-         * Safe to call even if the monitor already removed and completed the future.
-         */
-        void unregisterPermissionRequest(String requestId) {
-            Objects.requireNonNull(requestId, "requestId");
-            CompletableFuture<JsonNode> future;
-            synchronized (permissionLifecycleLock) {
-                future = pendingPermissions.remove(requestId);
-                if (globalPermissionRegistry != null) {
-                    globalPermissionRegistry.remove(requestId);
-                }
-            }
-            if (future != null && !future.isDone()) {
-                future.cancel(false);
-            }
-        }
-
-        /**
-         * Starts the background monitor thread that reads from the socket.
-         * Switches the channel to non-blocking mode for the duration of monitoring.
-         * Must be called before the agent loop starts.
-         */
-        void startMonitor() {
-            if (channel == null) {
-                log.debug("Cancel monitor: no socket channel available, skipping monitor");
-                return;
-            }
-            monitorThread = Thread.ofVirtual()
-                    .name("aceclaw-cancel-monitor")
-                    .start(this::monitorLoop);
-        }
-
-        /**
-         * Resolves every pending permission request as cancelled. Used by
-         * the {@code agent.cancel} branch of the monitor so the agent
-         * loop's {@code future.get()} in {@code PermissionAwareTool.execute}
-         * unblocks immediately with {@code CancellationException}
-         * instead of waiting up to {@link #PERMISSION_RESPONSE_TIMEOUT_MS}
-         * (120 s) — without this, Ctrl-C sets the cancellation token
-         * but the turn stays blocked on the permission future, so
-         * {@code stopMonitor()} never runs and the CLI's modal sits
-         * there until the daemon-side timeout fires.
-         *
-         * <p>Per request emits a best-effort {@code permission.cancelled}
-         * notification so the originating CLI's
-         * {@code TaskStreamReader.handlePermissionCancelled} can route
-         * it into {@link PermissionBridge#cancelExternal} — that path
-         * both completes the task-thread future and pokes the modal's
-         * polling loop so it dismisses without waiting on stdin.
-         * Send failures (socket closed, peer disconnected) don't roll
-         * back the future cancellation: the in-memory state must stay
-         * consistent even if the wire send fails.
-         *
-         * <p><b>Ordering:</b> the notification is sent <em>before</em>
-         * {@code future.cancel(false)}. Cancelling the future first
-         * would race the agent thread to the shared socket — once
-         * unblocked it returns a denied {@link ToolResult} and the
-         * turn's finally block emits the final JSON-RPC response. If
-         * that response wins the write, the CLI's
-         * {@code TaskStreamReader} stops reading further notifications
-         * and the trailing {@code permission.cancelled} is dropped on
-         * the floor, leaving the y/n modal stuck on stdin.
-         *
-         * <p>Idempotent. The snapshot + map drain runs under
-         * {@link #permissionLifecycleLock} so it's atomic against other
-         * writers (registerPermissionRequest / stopMonitor); the
-         * subsequent socket writes and future cancellations happen
-         * OUTSIDE the lock so a slow peer can't gate concurrent
-         * permission lifecycle operations.
-         */
-        private void cancelAllPendingPermissions(String via) {
-            // Drain the maps under the lock — but DON'T cancel futures yet.
-            // The agent loop's future.get() stays blocked, so it can't race
-            // us to the socket while we're sending permission.cancelled below.
-            var snapshot = new ArrayList<Map.Entry<String, CompletableFuture<JsonNode>>>();
-            synchronized (permissionLifecycleLock) {
-                pendingPermissions.forEach((rid, future) -> {
-                    snapshot.add(Map.entry(rid, future));
-                    if (globalPermissionRegistry != null) {
-                        // Match by future identity — same rationale as stopMonitor.
-                        var pending = globalPermissionRegistry.get(rid);
-                        if (pending != null && pending.future() == future) {
-                            globalPermissionRegistry.remove(rid, pending);
-                        }
-                    }
-                });
-                pendingPermissions.clear();
-            }
-            // Outside lock: send dismissal first, THEN cancel the future.
-            // See class-level javadoc above for why this ordering is required.
-            for (var entry : snapshot) {
-                // Skip futures already resolved by a concurrent path —
-                // typically a WS permission.response that completed the
-                // future via routePermissionResponse without taking
-                // permissionLifecycleLock (per-context pendingPermissions
-                // is only drained by the monitor's UDS-side handler and
-                // by agent-loop unregister, so we can snapshot an entry
-                // whose future was already approved). Emitting an
-                // approved=false cancellation here would advertise a
-                // denial for a request that actually won via approval —
-                // Codex P2 on PR #493. Tiny isDone-then-completed race
-                // remains (microseconds) but is bounded and the CLI's
-                // PermissionBridge.cancelExternal uses putIfAbsent so a
-                // late-stale dismissal is dropped on the receiving side.
-                var future = entry.getValue();
-                if (future.isDone()) {
-                    log.debug("Cancel monitor: skipping permission.cancelled for "
-                            + "already-resolved requestId={}", entry.getKey());
-                    continue;
-                }
-                try {
-                    var cancelParams = objectMapper.createObjectNode();
-                    cancelParams.put("requestId", entry.getKey());
-                    cancelParams.put("approved", false);
-                    cancelParams.put("via", via);
-                    delegate.sendNotification("permission.cancelled", cancelParams);
-                } catch (IOException e) {
-                    log.debug("Cancel monitor: failed to send permission.cancelled "
-                            + "for requestId={}: {}", entry.getKey(), e.getMessage());
-                }
-                future.cancel(false);
-            }
-        }
-
-        /**
-         * Stops the monitor thread cleanly without interrupting.
-         * Sets the stopped flag and wakes the selector so the thread exits promptly.
-         */
-        void stopMonitor() {
-            stopped = true;
-            // Cancel all pending permission futures so waiting threads unblock.
-            // No permission.cancelled fan-out here: stopMonitor runs at the
-            // request's finally block, by which point the turn is wrapping up
-            // and the CLI will see stream.cancelled / final response anyway.
-            // The agent.cancel branch (cancelAllPendingPermissions) is what
-            // handles the in-flight dismissal case.
-            synchronized (permissionLifecycleLock) {
-                pendingPermissions.forEach((rid, future) -> {
-                    future.cancel(false);
-                    if (globalPermissionRegistry != null) {
-                        // Drop only OUR entries from the daemon-wide registry —
-                        // other concurrent prompts may still have pending
-                        // permissions there. Iterating pendingPermissions
-                        // (per-context) gives us the correct subset.
-                        //
-                        // Match by future identity (not the wrapping
-                        // PendingPermission record): the registry value
-                        // wraps the same future together with sessionId
-                        // and context, so a naive remove(rid, future)
-                        // wouldn't equal the wrapper and would silently
-                        // leak stale entries. Look up first, then remove
-                        // by record-equality only when the contained
-                        // future matches.
-                        var pending = globalPermissionRegistry.get(rid);
-                        if (pending != null && pending.future() == future) {
-                            globalPermissionRegistry.remove(rid, pending);
-                        }
-                    }
-                });
-                pendingPermissions.clear();
-            }
-            var sel = selector;
-            if (sel != null) {
-                sel.wakeup();
-            }
-            var thread = monitorThread;
-            if (thread != null) {
-                try {
-                    thread.join(2000);
-                } catch (InterruptedException e) {
-                    Thread.currentThread().interrupt();
-                }
-                monitorThread = null;
-            }
-        }
-
-        private void monitorLoop() {
-            try {
-                channel.configureBlocking(false);
-                selector = Selector.open();
-                channel.register(selector, SelectionKey.OP_READ);
-
-                var buffer = ByteBuffer.allocate(READ_BUFFER_SIZE);
-
-                while (!stopped && !cancellationToken.isCancelled()) {
-                    int ready = selector.select(SELECT_TIMEOUT_MS);
-                    if (stopped || cancellationToken.isCancelled()) break;
-                    if (ready == 0) continue;
-
-                    selector.selectedKeys().clear();
-
-                    buffer.clear();
-                    int bytesRead = channel.read(buffer);
-                    if (bytesRead == -1) {
-                        log.debug("Cancel monitor: connection closed, triggering cancellation");
-                        cancellationToken.cancel();
-                        return;
-                    }
-                    if (bytesRead == 0) continue;
-
-                    buffer.flip();
-                    synchronized (lineBuilder) {
-                        lineBuilder.append(StandardCharsets.UTF_8.decode(buffer));
-
-                        // Process complete lines
-                        int newlineIdx;
-                        while ((newlineIdx = lineBuilder.indexOf("\n")) != -1) {
-                            var line = lineBuilder.substring(0, newlineIdx).trim();
-                            lineBuilder.delete(0, newlineIdx + 1);
-
-                            if (line.isEmpty()) continue;
-
-                            try {
-                                var message = objectMapper.readTree(line);
-                                String method = message.has("method")
-                                        ? message.get("method").asText("") : "";
-
-                                if ("agent.cancel".equals(method)) {
-                                    log.info("Cancel monitor: received agent.cancel");
-                                    cancellationToken.cancel();
-                                    // Unblock PermissionAwareTool.execute's future.get():
-                                    // setting the cancellation token alone doesn't propagate to
-                                    // CompletableFutures, so without this the turn sits blocked
-                                    // until PERMISSION_RESPONSE_TIMEOUT_MS (120 s) — and because
-                                    // the request never completes, stopMonitor() never runs and
-                                    // the CLI's y/n modal stays on screen waiting on stdin.
-                                    cancelAllPendingPermissions("agent-cancel");
-                                    return;
-                                } else if ("permission.response".equals(method)) {
-                                    log.debug("Cancel monitor: routing permission.response");
-                                    var respParams = message.get("params");
-                                    var rid = respParams != null && respParams.has("requestId")
-                                            ? respParams.get("requestId").asText() : null;
-                                    if (rid != null) {
-                                        var future = pendingPermissions.remove(rid);
-                                        if (future != null) {
-                                            // Keep the daemon-wide registry in sync —
-                                            // only when THIS context owned the request
-                                            // (local future found). Removing
-                                            // unconditionally would let a stale or
-                                            // cross-session CLI message delete another
-                                            // context's legitimate entry, causing the
-                                            // browser response for that request to be
-                                            // rejected as "unknown requestId" until
-                                            // timeout (codex P1 review on PR #454).
-                                            if (globalPermissionRegistry != null) {
-                                                globalPermissionRegistry.remove(rid);
-                                            }
-                                            future.complete(message);
-                                        } else {
-                                            log.warn("Cancel monitor: no pending request for requestId={}, dropping stale permission.response", rid);
-                                        }
-                                    } else {
-                                        log.warn("Cancel monitor: permission.response missing requestId, dropping message");
-                                    }
-                                } else if ("resume.response".equals(method)) {
-                                    log.debug("Cancel monitor: routing resume.response to fallback");
-                                    unmatchedResponses.offer(message);
-                                } else {
-                                    log.debug("Cancel monitor: ignoring '{}'", method);
-                                }
-                            } catch (Exception e) {
-                                log.warn("Cancel monitor: failed to parse message: {}", e.getMessage());
-                            }
-                        }
-                    }
-                }
-            } catch (IOException e) {
-                if (!stopped && !cancellationToken.isCancelled()) {
-                    log.debug("Cancel monitor: I/O error: {}", e.getMessage());
-                    cancellationToken.cancel();
-                }
-            } finally {
-                // Restore blocking mode so subsequent reads work normally
-                try {
-                    if (selector != null) selector.close();
-                    if (channel != null && channel.isOpen()) {
-                        channel.configureBlocking(true);
-                    }
-                } catch (IOException e) {
-                    log.warn("Cancel monitor: failed to restore blocking mode: {}", e.getMessage());
-                }
-                selector = null;
-            }
-        }
-
-        @Override
-        public void sendNotification(String method, Object params) throws IOException {
-            if ("stream.tool_use".equals(method)) {
-                toolUseCount.incrementAndGet();
-            }
-            delegate.sendNotification(method, params);
-        }
-
-        /**
-         * Reads the next unmatched response from the fallback queue.
-         * Used by ResumeRouter and other non-permission reads.
-         * Returns null if cancelled or the monitor thread is no longer running.
-         */
-        @Override
-        public JsonNode readMessage() throws IOException {
-            return readMessage(PERMISSION_RESPONSE_TIMEOUT_MS);
-        }
-
-        @Override
-        public JsonNode readMessage(long timeoutMs) throws IOException {
-            long deadline = System.currentTimeMillis() + timeoutMs;
-            while (!cancellationToken.isCancelled()) {
-                long now = System.currentTimeMillis();
-                long remaining = deadline - now;
-                if (remaining <= 0) {
-                    throw new SocketTimeoutException("response timeout after " + timeoutMs + "ms");
-                }
-                try {
-                    JsonNode msg = unmatchedResponses.poll(Math.min(500, remaining), TimeUnit.MILLISECONDS);
-                    if (msg != null) {
-                        return msg;
-                    }
-                    // Check if the monitor thread is still alive
-                    var thread = monitorThread;
-                    if (thread == null || !thread.isAlive()) {
-                        return null;
-                    }
-                } catch (InterruptedException e) {
-                    Thread.currentThread().interrupt();
-                    return null;
-                }
-            }
-            return null;
-        }
-    }
-
-    // -- Permission-aware tool wrapper -------------------------------------
-
-    /**
-     * Wraps a tool with permission checking. Before executing, checks
-     * the permission manager and, if needed, sends a permission request
-     * to the client and waits for the response.
-     */
-    private static final class PermissionAwareTool implements Tool, CancellationAware {
-
-        private final Tool delegate;
-        private final PermissionManager permissionManager;
-        private final CancelAwareStreamContext context;
-        private final ObjectMapper objectMapper;
-        private final HookExecutor hookExecutor;
-        private final String sessionId;
-        private final String cwd;
-        private final AntiPatternPreExecutionGate antiPatternGate;
-        private final java.util.function.Supplier<AntiPatternGateOverrideStatus> antiPatternOverrideSupplier;
-        private final AntiPatternGateFeedbackStore antiPatternGateFeedbackStore;
-        private final CandidateStore candidateStore;
-        private final RuntimeMetricsExporter metricsExporter;
-        // #480 PR 3: Provenance-chain inputs. {@code promptId} is per-request
-        // (constant across all tool calls in this prompt); the step-id
-        // supplier is read lazily on each tool call so it captures the plan
-        // step that's active at the moment of the permission check, not the
-        // step that was active when the registry was constructed.
-        private final dev.aceclaw.security.ids.PromptId promptId;
-        private final java.util.function.Supplier<String> currentStepIdSupplier;
-
-        PermissionAwareTool(Tool delegate, PermissionManager permissionManager,
-                            CancelAwareStreamContext context, ObjectMapper objectMapper,
-                            HookExecutor hookExecutor, String sessionId, String cwd,
-                            AntiPatternPreExecutionGate antiPatternGate,
-                            java.util.function.Supplier<AntiPatternGateOverrideStatus> antiPatternOverrideSupplier,
-                            AntiPatternGateFeedbackStore antiPatternGateFeedbackStore,
-                            CandidateStore candidateStore,
-                            RuntimeMetricsExporter metricsExporter,
-                            dev.aceclaw.security.ids.PromptId promptId,
-                            java.util.function.Supplier<String> currentStepIdSupplier) {
-            this.delegate = delegate;
-            this.permissionManager = permissionManager;
-            this.context = context;
-            this.objectMapper = objectMapper;
-            this.hookExecutor = hookExecutor;
-            this.sessionId = sessionId;
-            this.cwd = cwd;
-            this.antiPatternGate = antiPatternGate;
-            this.antiPatternOverrideSupplier = antiPatternOverrideSupplier;
-            this.antiPatternGateFeedbackStore = antiPatternGateFeedbackStore;
-            this.candidateStore = candidateStore;
-            this.metricsExporter = metricsExporter;
-            this.promptId = Objects.requireNonNull(promptId, "promptId");
-            this.currentStepIdSupplier = Objects.requireNonNull(currentStepIdSupplier, "currentStepIdSupplier");
-        }
-
-        @Override
-        public void setCancellationToken(CancellationToken token) {
-            if (delegate instanceof CancellationAware ca) {
-                ca.setCancellationToken(token);
-            }
-        }
-
-        @Override
-        public String name() {
-            return delegate.name();
-        }
-
-        @Override
-        public String description() {
-            return delegate.description();
-        }
-
-        @Override
-        public JsonNode inputSchema() {
-            return delegate.inputSchema();
-        }
-
-        @Override
-        public ToolResult execute(String inputJson) throws Exception {
-            // --- PreToolUse hooks (before permission check) ---
-            String effectiveInputJson = inputJson;
-            if (hookExecutor != null) {
-                try {
-                    var toolInput = objectMapper.readTree(inputJson);
-                    if (toolInput == null || !toolInput.isObject()) {
-                        toolInput = objectMapper.createObjectNode();
-                    }
-                    var preEvent = new HookEvent.PreToolUse(sessionId, cwd, delegate.name(), toolInput);
-                    var hookResult = hookExecutor.execute(preEvent);
-
-                    switch (hookResult) {
-                        case HookResult.Block blocked -> {
-                            log.info("Tool {} blocked by PreToolUse hook: {}", delegate.name(), blocked.reason());
-                            return new ToolResult("Blocked by hook: " + blocked.reason(), true);
-                        }
-                        case HookResult.Proceed proceed -> {
-                            if (proceed.updatedInput() != null) {
-                                effectiveInputJson = objectMapper.writeValueAsString(proceed.updatedInput());
-                                log.debug("Tool {} input modified by PreToolUse hook", delegate.name());
-                            }
-                        }
-                        case HookResult.Error err ->
-                                log.warn("PreToolUse hook error for {}: {}", delegate.name(), err.message());
-                    }
-                } catch (Exception e) {
-                    log.warn("PreToolUse hook failed for {}: {}", delegate.name(), e.getMessage());
-                }
-            }
-
-            // --- Permission check ---
-            // Determine the permission level for this tool
-            // MCP tools default to EXECUTE since they can have side effects
-            var level = delegate.name().startsWith("mcp__")
-                    ? PermissionLevel.EXECUTE
-                    : TOOL_PERMISSION_LEVELS.getOrDefault(delegate.name(), PermissionLevel.EXECUTE);
-
-            // Build a human-readable description of what the tool will do
-            var toolDescription = buildToolDescription(delegate.name(), effectiveInputJson);
-            final String finalInputJson = effectiveInputJson;
-
-            // #480 PR 2: routing the call through ToolPermissionRouter keeps
-            // the structured-vs-legacy branching logic in one testable place.
-            // CapabilityAware tools take the structured path; everything else
-            // hits the legacy PermissionRequest entry point. Both ultimately
-            // share PermissionManager's single decision/audit pipeline.
-            //
-            // #480 PR 3: build the full Provenance here so the structured
-            // path receives the actual rootPrompt and (when running inside a
-            // plan) the active planStepId. The supplier read lazily picks up
-            // whatever step is current at this exact moment — sequencing
-            // matters because a single agent.prompt can transition through
-            // multiple plan steps.
-            var stepIdRaw = currentStepIdSupplier.get();
-            var provenance = new dev.aceclaw.security.Provenance(
-                    java.util.Optional.of(promptId),
-                    java.util.Optional.of(new dev.aceclaw.security.ids.SessionId(sessionId)),
-                    stepIdRaw == null
-                            ? java.util.Optional.<dev.aceclaw.security.ids.PlanStepId>empty()
-                            : java.util.Optional.of(new dev.aceclaw.security.ids.PlanStepId(stepIdRaw)),
-                    0,                          // subAgentDepth — top-level main loop
-                    java.util.List.of()         // chain — populated piecemeal in follow-ups
-            );
-            var decision = ToolPermissionRouter.check(
-                    delegate, finalInputJson, provenance, toolDescription, level,
-                    permissionManager, objectMapper);
-            var overrideStatus = antiPatternOverrideSupplier != null
-                    ? antiPatternOverrideSupplier.get()
-                    : new AntiPatternGateOverrideStatus(sessionId, delegate.name(), false, 0L, "");
-            var evaluatedGateDecision = antiPatternGate != null
-                    ? antiPatternGate.evaluate(delegate.name(), finalInputJson, toolDescription)
-                    : AntiPatternPreExecutionGate.Decision.allow();
-            var antiPatternDecision = !overrideStatus.active()
-                    ? evaluatedGateDecision
-                    : AntiPatternPreExecutionGate.Decision.allow();
-
-            if (overrideStatus.active()) {
-                emitGateNotification("OVERRIDE", overrideStatus, evaluatedGateDecision);
-            }
-
-            if (antiPatternDecision.action() == AntiPatternPreExecutionGate.Action.BLOCK) {
-                emitGateNotification("BLOCK", overrideStatus, antiPatternDecision);
-                recordBlockedRule(antiPatternDecision.ruleId());
-                log.info("Anti-pattern gate blocked tool {} (ruleId={})",
-                        delegate.name(), antiPatternDecision.ruleId());
-                return new ToolResult(
-                        "Anti-pattern gate blocked execution: "
-                                + "{gate=anti_pattern_preexec, action=BLOCK, ruleId="
-                                + antiPatternDecision.ruleId()
-                                + ", reason=\"" + antiPatternDecision.reason() + "\""
-                                + ", fallback=\"" + antiPatternDecision.fallback() + "\"}",
-                        true);
-            }
-            if (antiPatternDecision.action() == AntiPatternPreExecutionGate.Action.PENALIZE) {
-                emitGateNotification("PENALIZE", overrideStatus, antiPatternDecision);
-                log.info("Anti-pattern gate penalized tool {} (ruleId={})",
-                        delegate.name(), antiPatternDecision.ruleId());
-            }
-
-            switch (decision) {
-                case PermissionDecision.Approved ignored -> {
-                    if (metricsExporter != null) metricsExporter.recordPermissionDecision(false);
-                    var result = executeWithPostHooks(finalInputJson);
-                    maybeRecordFalsePositiveAndRollback(overrideStatus, evaluatedGateDecision, result);
-                    return result;
-                }
-
-                case PermissionDecision.Denied denied -> {
-                    if (metricsExporter != null) metricsExporter.recordPermissionDecision(true);
-                    log.info("Tool {} denied: {}", delegate.name(), denied.reason());
-                    return new ToolResult("Permission denied: " + denied.reason(), true);
-                }
-
-                case PermissionDecision.NeedsUserApproval approval -> {
-                    // Send permission request to the client and await via per-request future
-                    var requestId = "perm-" + UUID.randomUUID();
-                    long timeoutMs = CancelAwareStreamContext.PERMISSION_RESPONSE_TIMEOUT_MS;
-
-                    var future = context.registerPermissionRequest(requestId, sessionId);
-                    if (future.isCancelled()) {
-                        return new ToolResult("Permission denied: request cancelled", true);
-                    }
-                    try {
-                        var params = objectMapper.createObjectNode();
-                        params.put("tool", delegate.name());
-                        params.put("description", approval.prompt());
-                        params.put("requestId", requestId);
-                        // Stamp the originating tool_use id so the
-                        // dashboard reducer can mark the SPECIFIC tool
-                        // node as awaiting (instead of falling back to
-                        // activeNodeId, which collapses onto whichever
-                        // parallel tool_use was emitted last). Without
-                        // this, parallel tool calls each fire their own
-                        // permission.request but only one node shows
-                        // the "click ✓/✗" chip — the user can only
-                        // approve one from the dashboard.
-                        var toolUseId =
-                                dev.aceclaw.core.agent.ToolExecutionContext.currentToolUseId();
-                        if (toolUseId != null) {
-                            params.put("toolUseId", toolUseId);
-                        }
-                        context.sendNotification("permission.request", params);
-
-                        // Each thread waits on its own future — no cross-delivery possible
-                        var responseMsg = future.get(timeoutMs, TimeUnit.MILLISECONDS);
-
-                        // Parse the permission response
-                        var responseParams = responseMsg.get("params");
-                        if (responseParams == null) {
-                            return new ToolResult("Permission denied: invalid response from client", true);
-                        }
-
-                        boolean approved = responseParams.has("approved")
-                                && responseParams.get("approved").asBoolean(false);
-                        boolean remember = responseParams.has("remember")
-                                && responseParams.get("remember").asBoolean(false);
-
-                        if (!approved) {
-                            if (metricsExporter != null) metricsExporter.recordPermissionDecision(true);
-                            log.info("Tool {} denied by user (requestId={})", delegate.name(), requestId);
-                            return new ToolResult("Permission denied by user", true);
-                        }
-                        if (metricsExporter != null) metricsExporter.recordPermissionDecision(false);
-
-                        // If user chose "remember", grant session-level approval.
-                        // Per-session scope (issue #456): the allow only applies
-                        // to THIS session — clicking Always Allow in workspace A
-                        // no longer silently disables the prompt in workspace B.
-                        if (remember) {
-                            permissionManager.approveForSession(sessionId, delegate.name());
-                        }
-
-                        log.info("Tool {} approved by user (requestId={}, remember={})",
-                                delegate.name(), requestId, remember);
-                        var result = executeWithPostHooks(finalInputJson);
-                        maybeRecordFalsePositiveAndRollback(overrideStatus, evaluatedGateDecision, result);
-                        return result;
-
-                    } catch (InterruptedException e) {
-                        Thread.currentThread().interrupt();
-                        if (metricsExporter != null) metricsExporter.recordPermissionDecision(true);
-                        log.info("Tool {} permission request interrupted (requestId={})",
-                                delegate.name(), requestId);
-                        return new ToolResult("Permission denied: request interrupted", true);
-                    } catch (TimeoutException e) {
-                        if (metricsExporter != null) metricsExporter.recordPermissionDecision(true);
-                        log.info("Tool {} permission response timed out (requestId={})",
-                                delegate.name(), requestId);
-                        long timeoutSeconds = TimeUnit.MILLISECONDS.toSeconds(timeoutMs);
-                        return new ToolResult(
-                                "Permission pending timeout: no response from client within "
-                                        + timeoutSeconds + "s", true);
-                    } catch (CancellationException e) {
-                        log.info("Tool {} permission request cancelled (requestId={})",
-                                delegate.name(), requestId);
-                        return new ToolResult("Permission denied: request cancelled", true);
-                    } catch (ExecutionException e) {
-                        log.error("Failed to receive permission response for tool {}: {}",
-                                delegate.name(), e.getMessage());
-                        return new ToolResult("Permission check failed: " + e.getMessage(), true);
-                    } catch (IOException e) {
-                        log.error("Failed to communicate permission request for tool {}: {}",
-                                delegate.name(), e.getMessage());
-                        return new ToolResult("Permission check failed: " + e.getMessage(), true);
-                    } finally {
-                        context.unregisterPermissionRequest(requestId);
-                    }
-                }
-            }
-        }
-
-        /**
-         * Executes the tool and fires PostToolUse or PostToolUseFailure hooks.
-         */
-        private ToolResult executeWithPostHooks(String inputJson) throws Exception {
-            ToolResult result;
-            try {
-                result = delegate.execute(inputJson);
-            } catch (Exception e) {
-                // Fire PostToolUseFailure hook (non-blocking, fire-and-forget)
-                var msg = e.getMessage() != null ? e.getMessage() : e.getClass().getSimpleName();
-                firePostHookAsync(inputJson, null, msg);
-                throw e;
-            }
-
-            // Fire PostToolUse or PostToolUseFailure based on result
-            if (result.isError()) {
-                var msg = result.output() != null ? result.output() : "Tool error";
-                firePostHookAsync(inputJson, null, msg);
-            } else {
-                firePostHookAsync(inputJson, result.output(), null);
-            }
-
-            return result;
-        }
-
-        /**
-         * Fires PostToolUse or PostToolUseFailure hooks on a virtual thread (fire-and-forget).
-         */
-        private void firePostHookAsync(String inputJson, String output, String error) {
-            if (hookExecutor == null) return;
-            Thread.ofVirtual().name("hook-post-" + delegate.name()).start(() -> {
-                try {
-                    var toolInput = objectMapper.readTree(inputJson);
-                    if (toolInput == null || !toolInput.isObject()) {
-                        toolInput = objectMapper.createObjectNode();
-                    }
-                    HookEvent event;
-                    if (error != null) {
-                        event = new HookEvent.PostToolUseFailure(
-                                sessionId, cwd, delegate.name(), toolInput, error);
-                    } else {
-                        event = new HookEvent.PostToolUse(
-                                sessionId, cwd, delegate.name(), toolInput,
-                                output != null ? output : "");
-                    }
-                    hookExecutor.execute(event);
-                } catch (Exception e) {
-                    log.warn("Post-tool hook failed for {}: {}", delegate.name(), e.getMessage());
-                }
-            });
-        }
-
-        /**
-         * Builds a human-readable description of what the tool will do based on its input.
-         */
-        private String buildToolDescription(String toolName, String inputJson) {
-            try {
-                var input = objectMapper.readTree(inputJson);
-                return switch (toolName) {
-                    case "bash" -> "Execute: " + (input.has("command")
-                            ? input.get("command").asText() : "(unknown command)");
-                    case "write_file" -> "Write to file: " + (input.has("file_path")
-                            ? input.get("file_path").asText() : "(unknown path)");
-                    case "edit_file" -> "Edit file: " + (input.has("file_path")
-                            ? input.get("file_path").asText() : "(unknown path)");
-                    case "read_file" -> "Read file: " + (input.has("file_path")
-                            ? input.get("file_path").asText() : "(unknown path)");
-                    case "glob" -> "Search files: " + (input.has("pattern")
-                            ? input.get("pattern").asText() : "(unknown pattern)");
-                    case "grep" -> "Search content: " + (input.has("pattern")
-                            ? input.get("pattern").asText() : "(unknown pattern)");
-                    case "list_directory" -> "List directory: " + (input.has("path")
-                            ? input.get("path").asText() : "(working directory)");
-                    case "web_fetch" -> "Fetch URL: " + (input.has("url")
-                            ? input.get("url").asText() : "(unknown url)");
-                    case "web_search" -> "Web search: " + (input.has("query")
-                            ? input.get("query").asText() : "(unknown query)");
-                    case "browser" -> "Browser " + (input.has("action")
-                            ? input.get("action").asText() : "(unknown action)") +
-                            (input.has("url") ? ": " + input.get("url").asText() : "");
-                    case "applescript" -> "Execute AppleScript (" +
-                            (input.has("script") ? input.get("script").asText().length() + " chars" : "unknown") + ")";
-                    case "screen_capture" -> "Capture screenshot" +
-                            (input.has("region") ? " (region: " + input.get("region").asText() + ")" : "");
-                    default -> {
-                        if (toolName.startsWith("mcp__")) {
-                            // MCP tools: show server and tool name
-                            var parts = toolName.split("__", 3);
-                            var server = parts.length > 1 ? parts[1] : "unknown";
-                            var tool = parts.length > 2 ? parts[2] : "unknown";
-                            yield "MCP [" + server + "] " + tool;
-                        } else {
-                            yield "Execute tool: " + toolName;
-                        }
-                    }
-                };
-            } catch (Exception e) {
-                return "Execute tool: " + toolName;
-            }
-        }
-
-        private void emitGateNotification(String action,
-                                          AntiPatternGateOverrideStatus overrideStatus,
-                                          AntiPatternPreExecutionGate.Decision gateDecision) {
-            try {
-                var params = objectMapper.createObjectNode();
-                params.put("sessionId", sessionId);
-                params.put("tool", delegate.name());
-                params.put("gate", "anti_pattern_preexec");
-                params.put("action", action);
-                if (gateDecision != null) {
-                    if (!gateDecision.ruleId().isBlank()) params.put("ruleId", gateDecision.ruleId());
-                    if (!gateDecision.reason().isBlank()) params.put("reason", gateDecision.reason());
-                    if (!gateDecision.fallback().isBlank()) params.put("fallback", gateDecision.fallback());
-                }
-                if (overrideStatus != null && overrideStatus.active()) {
-                    params.put("override", true);
-                    params.put("overrideTtlSeconds", overrideStatus.ttlSecondsRemaining());
-                    if (!overrideStatus.reason().isBlank()) {
-                        params.put("overrideReason", overrideStatus.reason());
-                    }
-                } else {
-                    params.put("override", false);
-                }
-                context.sendNotification("stream.gate", params);
-            } catch (Exception e) {
-                log.debug("Failed to emit stream.gate notification: {}", e.getMessage());
-            }
-        }
-
-        private void recordBlockedRule(String ruleId) {
-            if (antiPatternGateFeedbackStore == null || ruleId == null || ruleId.isBlank()) {
-                return;
-            }
-            antiPatternGateFeedbackStore.recordBlocked(ruleId);
-        }
-
-        private void maybeRecordFalsePositiveAndRollback(
-                AntiPatternGateOverrideStatus overrideStatus,
-                AntiPatternPreExecutionGate.Decision evaluatedGateDecision,
-                ToolResult toolResult) {
-            if (antiPatternGateFeedbackStore == null || toolResult == null || toolResult.isError()) {
-                return;
-            }
-            if (overrideStatus == null || !overrideStatus.active()) {
-                return;
-            }
-            if (evaluatedGateDecision == null || evaluatedGateDecision.action() != AntiPatternPreExecutionGate.Action.BLOCK) {
-                return;
-            }
-            String ruleId = evaluatedGateDecision.ruleId();
-            if (ruleId == null || ruleId.isBlank()) {
-                return;
-            }
-            boolean shouldRollback = antiPatternGateFeedbackStore.recordFalsePositive(ruleId);
-            if (shouldRollback) {
-                autoRollbackCandidateRule(ruleId);
-            }
-        }
-
-        private void autoRollbackCandidateRule(String ruleId) {
-            if (candidateStore == null || !ruleId.startsWith("candidate:")) {
-                return;
-            }
-            String candidateId = ruleId.substring("candidate:".length());
-            if (candidateId.isBlank()) {
-                return;
-            }
-            try {
-                var rollback = candidateStore.rollbackPromoted(
-                        candidateId, "AUTO_ROLLBACK_FALSE_POSITIVE_RATE");
-                if (rollback.isPresent()) {
-                    log.info("Auto-rolled back anti-pattern candidate due to false-positive gate rate: {}",
-                            candidateId);
-                }
-            } catch (Exception e) {
-                log.warn("Failed auto rollback for candidate {}: {}", candidateId, e.getMessage());
-            }
-        }
-    }
 
     private record AntiPatternGateOverride(Instant expiresAt, String reason) {}
 }
